+++ client/Packages/com.beamable/Runtime/Player/PlayerInventory.cs
@@ -462,6 +462,9 @@ namespace Beamable.Player
 			return group[0];
 		}
 
+		/// <inheritdoc cref="GetCurrency(Beamable.Common.Inventory.CurrencyRef)"/>
+		public PlayerCurrency GetCurrency(string currencyRef) => GetCurrency(new CurrencyRef(currencyRef));
+
 		/// <summary>
 		/// Get a category of <see cref="PlayerItem"/> for a given type.
 		/// If you have subtypes of <see cref="ItemContent"/>,
@@ -483,6 +486,23 @@ namespace Beamable.Player
 			return itemGroup;
 		}
 
+		/// <inheritdoc cref="GetItems(Beamable.Common.Inventory.ItemRef)"/>
+		public PlayerItemGroup GetItems(string itemRef) => GetItems(new ItemRef(itemRef));
+
+		/// <inheritdoc cref="GetItems(Beamable.Common.Inventory.ItemRef)"/>
+		/// This method will is similar to <see cref="GetItems(Beamable.Common.Inventory.ItemRef)"/>, however it will make sure that the
+		/// <see cref="PlayerItemGroup"/> has been refreshed before completing the resulting promise.
+		/// <returns> A <see cref="Promise"/> containing the updated <see cref="PlayerItemGroup"/></returns>
+		public async Promise<PlayerItemGroup> LoadItems(ItemRef itemRef=null)
+		{
+			var items = GetItems(itemRef);
+			await items.Refresh();
+			return items;
+		}
+
+		/// <inheritdoc cref="LoadItems(Beamable.Common.Inventory.ItemRef)"/>
+		public Promise<PlayerItemGroup> LoadItems(string itemRef) => LoadItems(new ItemRef(itemRef));
+		
 		/// <summary>
 		/// Get a category of <see cref="PlayerCurrency"/> for a given type.
 		/// If you have subtypes of <see cref="CurrencyRef"/>,
@@ -503,6 +523,23 @@ namespace Beamable.Player
 			_saveHandle?.Save();
 			return currencyGroup;
 		}
+
+		/// <inheritdoc cref="GetCurrencies(Beamable.Common.Inventory.CurrencyRef)"/>
+		public PlayerCurrencyGroup GetCurrencies(string currencyRef) => GetCurrencies(new CurrencyRef(currencyRef));
+
+		/// <inheritdoc cref="GetCurrencies(Beamable.Common.Inventory.CurrencyRef)"/>
+		/// This method will is similar to <see cref="GetCurrencies(Beamable.Common.Inventory.CurrencyRef)"/>, however it will make sure that the
+		/// <see cref="PlayerCurrencyGroup"/> has been refreshed before completing the resulting promise.
+		/// <returns> A <see cref="Promise"/> containing the updated <see cref="PlayerCurrencyGroup"/></returns>
+		public async Promise<PlayerCurrencyGroup> LoadCurrencies(CurrencyRef currencyRef=null)
+		{
+			var currencies = GetCurrencies(currencyRef);
+			await currencies.Refresh();
+			return currencies;
+		}
+		
+		/// <inheritdoc cref="LoadCurrencies(Beamable.Common.Inventory.CurrencyRef)"/>
+		public Promise<PlayerCurrencyGroup> LoadCurrencies(string currencyRef) => LoadCurrencies(new CurrencyRef(currencyRef));
 
 
 		/// <summary>